+++ lib/davinci_dtr_test_kit/client_groups/dinner_adaptive/dtr_full_ehr_adaptive_dinner_questionnaire_workflow_group.rb
@@ -1,9 +1,9 @@
-require_relative '../full_ehr/dtr_full_ehr_adaptive_questionnaire_initial_retrieval_group'
+require_relative 'dtr_full_ehr_adaptive_questionnaire_initial_retrieval_group'
 require_relative '../full_ehr/dtr_full_ehr_prepopulation_attestation_test'
 require_relative '../full_ehr/dtr_full_ehr_prepopulation_override_attestation_test'
 require_relative '../full_ehr/dtr_full_ehr_saving_questionnaire_response_group'
-require_relative '../shared/dtr_adaptive_questionnaire_followup_questions_group'
-require_relative '../shared/dtr_adaptive_questionnaire_completion_group'
+require_relative 'dtr_adaptive_questionnaire_followup_questions_group'
+require_relative 'dtr_adaptive_questionnaire_completion_group'
 
 module DaVinciDTRTestKit
   class DTRFullEHRAdaptiveDinnerQuestionnaireWorkflowGroup < Inferno::TestGroup
